fix(views): reslove name collision in imports & rename submodules
