+++ src/administrator/views/__init__.py
@@ -1,8 +1,8 @@
-from .overview import overview
-from .settings import settings
-from .activity_and_history import activity_and_history
-from .users_and_access import users_and_access
-from .issues_and_broadcasts import issues_and_broadcasts
+from ._overview import overview
+from ._settings import settings
+from ._activity_and_history import activity_and_history
+from ._users_and_access import users_and_access
+from ._issues_and_broadcasts import issues_and_broadcasts
 
 __all__ = [
     "overview",
