+++ src/administrator/views/_activity_and_history.py
